Доделал utc offset, сейчас правилно отображаются. С сервера приходят неверные min и max temp для дня. Нужно брать из почасового прогноза

+++ app/src/main/java/com/example/weather/models/main/WeatherUiModel.kt
@@ -29,7 +29,7 @@ data class WeatherUiModel(
                     if (index == 0) null
                     else HourlyData(
                         weather.temp,
-                        getHour(Date(weather.time * 1000)),
+                        getHour(Date(weather.time * 1000)).toFloat(),
                         weather.iconId
                     )
                 }
